Add the attractor parameters as an effective subtitle
Since Axis3 does not have a subtitle attribute this needs to be
jury-rigged using line breaks. The issue is that the subtitle shares the
same font style as the title and Unicode needs to be used so the result
doesn't look great. I tried using LaTeXStrings, and I could apply
separate formatting to the title and the paremeters, but I couldn't
figure out a way to add line breaks so I couldn't get them on a new line.

I'm commit this in case I want to go back to this solution, but I think
I'm going to go with a separate label for each parameter in a GridLayout
on the side panel.