Generalize
I plan to add more attractors so this is the switch to more generic
programming. I will change the name of the package, probably to
something like ChaoticStrangeAttractors or something; although that does
seem particularly long it avoids clashing with other existing packages.
In accordance with this the `_Rossler` type is temporary for the time
being.

The `evolve!` macro is going to be particularly useful in reducing the
amount of boilerplate and making things easier to manage for a large(r)
number of attractors since the evolution code is going to be common to
all of them given that the three derivatives are and that particular
code carries with it a heavy number of parameters. Yes, it's an
unhygienic macro, but that's what makes it quite useful in this case.