Increase the default resolution for dt
So given that we're linearly interpolating the positions and the
evolution is based upon a timer, the value for dt does actually affect
the path since of course the system is chaotic and it's sensitive to
small variances. Decreasing the timer's interval is not a solution as
already the updating is quite laggy if the plot is interacted with and
we're pushing quite a number of lines which appears to consume
significant resources; 50 ms resolution (corresponding to 20 fps) seems
ideal for the timer, the lagginess can be dealt with by shifting around
the computational expensive operations off of the timer.

So, fix the timer, but decrease dt. This means each time the timer fires
for dt = 1 ms the attractor will loop through 50 times and then on
update the 51 element vectors containing each point evolved for each
coordinate through-out the interval is sent to `lines!` for rendering.

In practice this seems to improve performance and leads to expected
behaviour for the attractor.