Add a time field to the `Attractor` structs
I wanted to make this dynamic by making it an `Observable` and having
the `Label` update, but that slowed things down quite a bit, and the
evolution is already slow enough when the plot becomes sufficiently
populated.

I've changed the way the type is displayed as I'm planning on removing
the labels outright in the next commit since they clutter up the plot
and the code and the attractor can just be queried for the parameters.
For now they will display by default.