Improve performance by reducing the number of plotted objects
On every update (50 ms interval) line plots were being pushed onto the
axis and this was causing issues for some reason, probably having to do
with some Makie internal mechanism iterating through all of the plot
objects as it kept track of things so the system would get bogged down
rather quickly.

The solution is to simply keep track of all of the path points ourselves
and swap out the line plot on every update. The position points will get
quite numerous quickly as well for small dt values (every second 1000
new line points are added to the plot on the default setting), but this
seems to bear little relative performance impact compared to the
previous approach because in the end we're dealing with only one actual
composite line plot. It could also likely be useful having these path
points outside of the visualization aspect of the package since they
represent and quantify the system's evolution numerically.