Overload `show` instead of `display`
`display` (and even `print`) ultimately call `show` so it's better to
overload that. I think I was overloading `display` before doing any
of the custom printing; if returning nothing `show` adds an extra line
and since the REPL invokes `display` it's more direct in a sense and
avoids that extra line.