v0.1.0
So I'm finally releasing this since it works fairly well now.
It's gone through multiple iterations and even a rename and could have
stepped through multiple versions, but it didn't really make sense to
increment the version number while I was still actively working on a
project that was constantly changing and going through redesigns and
refactors; maybe it should make more sense in that regard, since this
has fairly matured since the initial, but if every commit practically
introduced a new version then there's no point in bothering.

Anyway, still can't get mp4 or webm to work on the README, the only way
I can even get it to embed / display is to use GitHub's drag & drop or
selection feature to upload the file as an asset, but even then there
seems to be no way to control it through HTML so no autoplay, no loop,
and the controls are visible; so it's better to just go with gif.

For actual encoding though, mp4 seems to be the best bet: it's higher in
quality, smaller in size, and encodes faster compared to webm with how
GLMakie hooks into FFmpeg to carry out the encoding.