Redesign `show` for efficiency
Since running multiple attractors is quite resource intensive and we
loop through fields for display it's better to separate things
especially since for vectors the 3-arg show is being tried and so also
has go through the display logic as well as constructing print streams
that will not be used.

:typeinfo was chosen for the IOContext branching check because :compact
returns false both for a single attractor and a vector of attractors,
and :limit returns true for both.

The idea is to automatically suppress printing if multiple paths are
being traced, as printing is slow as everything gets (re)-initialized.