Remove the AttractorSet type, make some timer & pause fixes
The extra meta type was making things more difficult to manage and
reason about and we really don't need an extra layer of abstraction.

For the fixes:

* pause is now passed at start_timers (formerly set_timers) as not doing
so requires an extra interaction in order to flip the pause state if the
second timer fires;

* the figure and the pause Observable is now explicitly reinitialized
since passing attract! an initialized Attractor would retain the
registered listeners from the previous call.