Don't overload `display` for `Vector{<:Attractor}`; return nothing
Okay so since the Julia manual explicitly recommends against overloading
containers like this and given that we're suppressing its output
completely under attraction then maybe it's best to go with this.

All `attract!` methods now return nothing, which is probably more
sensible given that most of time you're constructing the attractor in
the first place.

That being said it's nice calling the function with a comprehension
using the default and only varying the initial conditions without pre-
constructing the attractor and having it returned so I might revert this
anyway; I'll have to think about it some more.