Fix a hang if an initialized attractor is passed to `attract!`
Had to bisect this and it still wasn't particularly obvious.
Basically, each individual position vector was being sent one-by-one
to the plotting machine and given that the number of points accumulate
rapidly this was computationally taxing which caused the program to
essentially hang on an undefined `lines!` method; obviously this doesn't
happen when passing the entire vector of `Point3`s at once.