Use an explicit midpoint method to solve the system
The time evolution had previously been implented using a naïve type of
Euler method to solve the differential equations which although fast is
comparatively inaccurate.

Also, the evolution is now carried out using a standard function rather
than a macro and the attractor functors now return the derivatives at a
given set of coordinates which is a much more intuitive implementation.

The time step has been doubled as the evolution now has twice the number
of function calls; given that the midpoint method is much more accurate
even with double the time step this should be an improvement and the
speed should be properly compensated for after reducing the number of
iterations per update interval.

It might be worth it to bring in some Runge-Kutta package at least for
the instantiation calls, but for now I want to keep a more minimal
package rather than bring in an extra dependency. This should be
accurate enough and it's simple enough to implement and maintain.

Anyway, this might improve stability as some attractors were broken for
some initial conditions.