Implement the classical 4th order Runge-Kutta method instead
After tuning the default time step and benchmarking this new method it's
definitely fast enough, even the spikes are well under the 50 ms update
interval. It's also simple enough to implement. So it's the
default now.

As for the accuracy, the local truncation error is O(dt^5) and the
global truncation error is O(dt^4) which should be good enough for most
short evolutions; the time step can always be varied at the attractor if
more accuracy is required, but in order to get the error on the order of
the machine epsilon for 64 bit floats `dt` needs to be an order of
magnitude smaller which leads to hundreds of `evolve!` calls and over a
thousand `attractor!` calls as well as many more trajectory points
having to be inventoried; this larger catalog being pushed simply makes
things much slower.

Maybe I should still bring in DifferentialEquations.jl from SciML as
they'll likely have more efficient and accurate methods, but again for
now this is good and it avoids bringing in what is possibly a heavy
dependency.